refactor(utils): extract format_number_human to num_human module
Add configurable HumanFormat with builder pattern supporting:
- Grouping styles (Standard, Indian, None)
- Custom separator character
- Compact notation (K/M/B/T suffixes)
- Configurable threshold and decimal places

+++ packages/nextclade/src/align/seed_alignment.rs
@@ -292,7 +292,7 @@ mod tests {
     let qry_len = 30;
     let ref_len = 40;
 
-    let _band: (Vec<Stripe>, usize) = create_alignment_band(
+    let _band: (Vec<Stripe>, u64) = create_alignment_band(
       &seed_matches,
       qry_len,
       ref_len,
